fix scala test
